requirements_simple.txt
PyGithub>=1.59.0
GitPython>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0
tqdm>=4.66.0
aiohttp>=3.9.0
